+++ tests/src/secure/tests_ubi_secure_attach.c
@@ -51,8 +51,7 @@ static void *ztest_suite_setup(void)
 static void ztest_suite_before(void *ctx)
 {
 	(void)ctx;
-	ubi_test_partition_force_release_all();
-	zassert_ok(flash_erase(UBI_PARTITION_DEVICE, UBI_PARTITION_OFFSET, UBI_PARTITION_SIZE));
+	ubi_test_secure_before_impl();
 }
 
 /* Module interface function definitions -------------------------------------------------------- */